% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatedData.R
\name{imageToPatches}
\alias{imageToPatches}
\title{imageToPatches}
\usage{
imageToPatches(image, mask = NA, radius = 5, npatches = NA,
  groundTruth = 0)
}
\arguments{
\item{image}{a reference image defining the image space}

\item{mask}{mask for the image}

\item{radius}{a scalar defining the patch size}

\item{npatches}{number of patches to regular sample from the image}

\item{groundTruth}{single scalar or vector (size of mask) with groundTruth}
}
\value{
list containing the image patches and separately the data frame
holding the patch coordinate and the ground truth.
}
\description{
Convert an input image to a set of patches of given radius.  Return the local
patch position for each patch. Optionally return the ground truth data
associated with each patch, if available.
}
\examples{

rimg = makeImage( c(25,25) )
sim = simulateSphereData( rimg, radius = c( 2, 3 )  )
ptch = imageToPatches( sim$image, rimg * 0 + 1  )

}
\author{
Avants BB
}

