% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatedData.R
\name{simulateSphereData}
\alias{simulateSphereData}
\title{simulateSphereData}
\usage{
simulateSphereData(referenceImage, radius = c(5, 6), noiseLevel = c(0, 0.1),
  positionNoiseLevel = c(0, 2))
}
\arguments{
\item{referenceImage}{a reference image defining the image space}

\item{radius}{a two vector defining the radii of the inner and outer shells}

\item{noiseLevel}{a two vector defining the noise level}

\item{positionNoiseLevel}{a two vector defining the noise level controlling
how much the image moves randomly away from the center.}
}
\value{
list containing the single channel image, the multi-channel image,
the ground truth labels and the center of mass stored in a data frame.
}
\description{
simulate 2D or 3D image geometry and segmentation data. Return the single
channel image, the multi-channel image, center of mass and ground truth labels.
}
\examples{

rimg = makeImage( c(25,25) )
sim = simulateSphereData( rimg, radius = c( 2, 3 )  )

}
\author{
Avants BB
}

