% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translationCode.R
\name{writeANTsImageToNumpy}
\alias{writeANTsImageToNumpy}
\title{writeANTsImageToNumpy}
\usage{
writeANTsImageToNumpy(img, fn)
}
\arguments{
\item{img}{input antsImage}

\item{fn}{output filename (probably a .npy extension)}
}
\description{
Write antsImage type to numpy array.  This will produce a vector for a single
channel image and a matrix for a multi-channel image.  We reorder the data
such that the ordering of indices matches across the two systems.
}
\examples{

img = ANTsR::antsImageRead( ANTsR::getANTsRData( "r16" ) )
ofn = tempfile( fileext=".npy" )
writeANTsImageToNumpy( img, ofn )
img = ANTsR::makeImage( c( 5, 6, 7 ) )
writeANTsImageToNumpy( img, ofn )

# to read in python, do
# from PIL import Image
# import numpy as np
# n = 256
# data = np.load( ofn )
# array = np.reshape( data, [n, n] )
# from scipy.misc import toimage
# toimage(data).show()

}
\author{
Avants BB
}

